Converted CSEvents to class
